# packages/contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sophia application API",
    "version": "0.4.0-design",
    "description": "Sophia-owned binding contract. Not the Omnigent OpenAPI and not a claim that these routes are deployed."
  },
  "paths": {
...
    "/api/v1/rooms/{roomId}/exchanges": {
      "post": {
        "operationId": "startExchange",
        "summary": "Open the room's one exchange with Sophia against the current room revision. Any member may open it (a viewer's exchange can converse and read, never start or control work); it is refused while a guest is in the room. Joining the room never opens it.",
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "supabaseBearer": []
          }
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Accepted response; semantic delivery state must be inspected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeReceipt"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Not permitted"
          },
          "409": {
            "description": "Stale target, idempotency mismatch or unresolved state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request fails the declared schema or semantic field constraints",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependent service unavailable; reconcile writes before retry"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeRequest"
              }
            }
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A"
      }
    },
...
    "/api/v1/exchanges/{exchangeId}/look": {
      "post": {
        "operationId": "showSophia",
        "summary": "Show Sophia your own screen or camera, explicitly: a new observation epoch for that one source. Only the person sharing it can select it.",
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "supabaseBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exchangeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The exchange as it is now",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeState"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Not a member of this room's project, or not the person sharing this source"
          },
          "409": {
            "description": "The exchange is not in a state that allows this (ended, paused for a guest, or another person's floor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request fails the declared schema or semantic field constraints",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "The media service needed to confirm this is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/exchanges/{exchangeId}/resume": {
      "post": {
        "operationId": "resumeExchange",
        "summary": "Resume a paused exchange by explicit member action: after a guest left (the room's trusted presence must be member-only) or after the holder left and came back.",
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "supabaseBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exchangeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The exchange as it is now",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeState"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Not a member of this room's project, or not the person sharing this source"
          },
          "409": {
            "description": "The exchange is not in a state that allows this (ended, paused for a guest, or another person's floor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Request fails the declared schema or semantic field constraints",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "The media service needed to confirm this is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-sophia-implementation": "contract_only",
        "x-sophia-release-goal": "S1-05A"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "supabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "runtimeCapability": {
        "type": "http",
        "scheme": "bearer",
        "description": "A runtime capability: an opaque secret held by the execution host and stored by the service only as SHA-256. It is bound to one project, executor resource and runtime unit, and it is not a member identity."
      },
      "mediaBridgeCapability": {
        "type": "http",
        "scheme": "bearer",
        "description": "The media bridge's capability: an opaque secret on the bridge host; the API holds only its SHA-256. It names no member and grants no member authority: tool calls act for the speaker the exchange's input epoch binds, under that speaker's own role."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "retry": {
            "type": "string",
            "enum": [
              "safe_read",
              "same_admission_key",
              "reconcile_first",
              "reauthorize",
              "never"
            ]
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "retry"
        ]
      },
...
      "ExchangeRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "expectedRoomRevision": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "allowVision": {
            "type": "boolean"
          }
        },
        "required": [
          "expectedRoomRevision",
          "allowVision"
        ]
      },
      "ExchangeReceipt": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "exchangeId": {
            "type": "string",
            "format": "uuid"
          },
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "inputActorId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "exchangeId",
          "roomId",
          "revision",
          "inputActorId"
        ]
      },
...
      "LookRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "screen",
              "camera"
            ]
          }
        },
        "required": [
          "source"
        ]
      },
      "ExchangeState": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "exchangeId": {
            "type": "string",
            "format": "uuid"
          },
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "open",
              "paused",
              "ended"
            ]
          },
          "pauseReason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "guest",
                  "holder_left"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "inputEpoch": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "playbackEpoch": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "observationEpoch": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "revision": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "exchangeId",
          "roomId",
          "state",
          "pauseReason",
          "inputEpoch",
          "playbackEpoch",
          "observationEpoch",
          "revision"
        ]
      },
...
    }
  }
}